requirements.txt
torch>=1.9.1+cu111
torchvision==torchvision==0.10.1+cu111
torchaudio>=0.9.1 
torchmetrics==0.3.2
pytorch_lightning==1.3.4
opencv-python==4.6.0.66
tensorboard==2.11.2
scikit-image
numpy
SimpleITK>=2.1.1
PyYAML
pynrrd
tqdm
sklearn
filelock
nibabel
pydicom
medpy
setuptools==59.5.0
matplotlib
torchsummary
batchgenerators